// src/app/modules/configuration/general/models/event-type-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Model to add and update an invoice adjustment type
 */
export interface EventTypeView { 
    /**
     * Gets or Sets Name
     */
    name?: string;
    /**
     * Gets or Sets IsPositiveAdjustment
     */
    isCompetition?: number;
}